CMakePM: Parse the CMakeLists.txt file to find out the project's name
The displayName of the project is used for the macro %{Project:Name}.

We used to get the project's parent directory as the project's name,
until the CMake project was parsed and the real project name was set.

This way we can have the proper name from the start.

QMake doesn't have this problem, since the project name is part of the
project filename e.g. <project-name>.pro

Fixes: QTCREATORBUG-32996
Change-Id: I51b379b3db864f454260e39d6ad61bf2e52f9f51
Reviewed-by: Eike Ziller <eike.ziller@qt.io>