LanguageClient: Avoid accessing deleted hover handler
The hover handler is a member of the Client. As soon as the client is
deleted the hover handler is gone as well. As a result we need to ensure
to remove the hover handler from all editor widgets that previously got
a pointer to the hover handler. And since it is possible to open
documents and editors outside of the coreplugin we cannot rely on the
editor for document functionality of the coreplugin. As a result we have
to manually track open editor widgets and correctly unset the hover
handler on client destruction.

Fixes: QTCREATORBUG-32945
Change-Id: If0bd4d1470426725d7d43bb7c4f6be519effcbe9
Reviewed-by: Christian Stenger <christian.stenger@qt.io>