ProjectExplorer: Merge targetsettingspanel.* into projectwindow.cpp
The previous re-work of the main window (245965223450e69 etc from 2016)
was aiming at having project related data in loosely (or not at all)
coupled items inside a single big model displayed by a single tree.

The idea was that over time, new types of items would be needed
and could be added more easily if at least child items did not
explicitly need to know their parents types.

Part of the price for that was that signalling changes up needed some
kind of type-erasing mechanism, for which QAbstractItemModel::setData
with custom roles and data were used, which let to not-so-obvious
implicit dependendencies between the components.

Turned out that we didn't need that often (only additions were the
"Vanished Targets" where it 'worked',  and the "Hide inactive kits"
mechanism, which already did not fit into the system as it was only
not intrinsic part of the model, only there to have it displayable
by the view).

The recent split of the view solves the "Hide inactive kits" problem,
and with the knowledge that we don't have to extend the model very
often in practice, we can paddle back and drop the "type erasure"
overengineering and  simply let the children know the type of their
parents. Price for that before the actual dropping of the setData()
mechanism we need to either expose the parent items in the .h files,
or, since all relevant items are only needed in targetsettingspanel.cpp
and projectwindow.cpp, merge the file pairs.

This patch here does the latter,

Change-Id: Iad1b71048d857c8899807fadc9582b5732a5f0e4
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>