Revert "ES: Fix that IsoIconBrowser indirectly enabled QmlDesigner"
This reverts commit f8cfa50081c6c7890682833e92b64b4c3ac62b79.

QmlDesigner is disabled by default, and IsoIconBrowser is enabled by
default, and the goal was to prevent IsoIconBrowser from loading if
QmlDesigner isn't explicitly enabled.

Unfortunately this conflicts with what we do with AiAssistant:
LuaLanguageClient is disabled by default, and AiAssistant is enabled by
default, and we want AiAssistant to load and indirectly enable
LuaLanguageClient.

We also cannot "manually" enable LuaLanguageClient when the user
installs AiAssistant, because that would not fix any existing
installations.

Task-number: QTCREATORBUG-32907
Fixes: QTCREATORBUG-33010
Change-Id: I0bc95ee8c75804c45cae9e84a3c78d54a5c9e984
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>