ProjectExplorer: Unify ComboBoxModel and ProjectsModel
The model used in the project selector combobox was a kind of
wrapper around all actual project data, and the project model
run in a special mode with only one child of the root node
which was dynamically choosen via the active item in the
combobox model.

The reason for this setup are forgotten, unifying both models
makes the code simpler.

Change-Id: I4c3750ee68e47b3030f520b6551c5e9bbb7d0296
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>