CMakePM: Replace FileSystemWatcher with FilePathWatcher
Even that the CMake Project Manager watches only one file, this can
cause
freezes since when the file vanishes (the index file is replaced by a new
one) the FileSystemWatcher starts watching the directory, which leads to
the macOS qfilesystemwatcher_fsevents implementation going into a
QDirIterator frenzy iterating the watched directory (in the case of e.g.
Qt some 9000+ files).

Fixes: QTCREATORBUG-33191
Change-Id: Ic4355ffefe9e48fcff91a91c7753ee4272fd3d99
Reviewed-by: Eike Ziller <eike.ziller@qt.io>