ScxmlEditor: Fix transition item's target point computation
We need to take the bounding box offset into account when computing
target point using the target factor. The offset is used when mapping
to the target factor the target factor, we also need to use it when
mapping back from the target factor.

This fixes an issue where the computed target point would be off due to
the math being based on differerent rects (with or without offset). This
resulted in the transition end points jumping away from the cursor when
released.

Task-number: QTCREATORBUG-32654
Change-Id: I7f6d4cb149f203b9cbc19c3b8a3af551d4413eb2
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>