ScxmlEditor: Respect snap when moving transitions
When grabbing the end of a transition and moving it, the point will
snap to certain points to help align objects. This snap needs to be
taken into account when the grabber is released and we reposition the
end point a last time. Previously, we would simply use the current
mouse position, which could be off from the snapped to point.

Fixes: QTCREATORBUG-32654
Change-Id: I3e8f0d4027ba4586e4d8546215b647bd05e58eaf
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>