Debugger: Avoid requesting unusable amount of disassembler
Limit the requested range of dissassembler to a sensible amount when
trying to match the beginning of the requested range to the closest
function. Especially if there is no debug information for the debugged
binary the next known function might be completely unrelated and to far
away, resulting in a huge disassemble output that is not very user
friendly and potentially freezes or even crashes Qt Creator.

Fixes: QTCREATORBUG-32764
Change-Id: I6173daa4aeb0a551f51c4a4ea47e4ce7acae2363
Reviewed-by: hjk <hjk@qt.io>