CMakePM: Fix qmake detection on Windows for CMake import builds
The cross-compiled builds of Qt seem to deliver both qmake and qmake.bat
as scripts that point to the host qmake binary.

The qmake detection code would look after "qmake qmake.bat" and since
both are present "qmake" would be picked.

This broke the Qt detection on Windows for Android import builds. This
includes CMake presets.

Change-Id: I7a0272e721a19926bc58cd50282a3f6de274a05e
Reviewed-by: Cristian Adam <cristian.adam@qt.io>