CMake: Use configure_package_config_file for QtCreatorConfig
Move the contents of the QtCreatorConfig.cmake file to a cmake.in
template file, which will be instantiated by
configure_package_config_file.

${var} style references are replaced with @var@ style references,
and backslash escapes are removed.

Use two separate calls to configure_package_config_file, one to create
a build-tree config file, and one to create an install-tree config
file. The build-tree one will be used by the super projects that call
add_subdirectory(qtcreator), in which case the install prefix is the
project binary dir.
The install-tree is meant to be installed and will be used by regular
non-super projects.

Aside from better readability due to lack of escaping, a follow up
change will introduce a new exported variable which will contain the
relative path to the install prefix from the location of the Config
file.
configure_package_config_file allows us to compute
that path automatically.

Change-Id: I6899221f628af03a5e640f70c297001b8d59cc4c
Reviewed-by: Eike Ziller <eike.ziller@qt.io>