Android: Don't copy gradle files into the project
In seemingly random cases, the manifest/templates wizard copied the
Gradle files into the project directory although the checkbox said not
to do so.

As fix, this change initializes the boolean of the wizard (to false) and
sets the checked state in the wizard page based on that.

Amends: 1daa7d34a2f94d939b652f8f5013f55e41d3c006

Task-number: QTCREATORBUG-26476
Change-Id: I5386644312bcee971d76b9e9230e1df20cd08446
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>