QmlDesigner: Improve common type cache
const char* template parameter can trigger a string lenght
computation. MSVC seems not to optimize it. StaticString is
saving the size of the string. So it is avoiding it. It has
the advantage too of being more stable because the pointer
can be different in different DLLs.

The typeId is now computed for every synchronization. That
makes the synchronization slightly slower, but symplifies
the type getter.

Change-Id: Id7cf69e380bf56beaec595589cc1252e342473a0
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>