Axivion: Fix handling of local dashboard table
When switching from global to local dashboard we performed
slightly different which on the one hand led to displaying
no issues although there are some present.
On the other hand this broke the disabling of "special"
version items when using the local dashboard as we accesed
the project info too early.
Synchronize the local and the global approach and perform
the same way to avoid these issues.

Fixes: QTCREATORBUG-33012
Change-Id: I1aff2642657221ad76effc02780ba20386b5fa09
Reviewed-by: Mohammad Mehdi Salem Naraghi <mehdi.salem@qt.io>
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>