Centralize Settings storage a bit
Until now, the plugin manager effectively owned the settings, with
forwarded access from ICore and three places in Utils getting
settings injected.

This here moves the storage to Utils, removing the need for
injection, as all users depend on Utils.

Syntax-wise this changes access to use references instead of pointers.

Change-Id: If63258fbe37da2ad3dcf78f8c2eedb61753e7cc8
Reviewed-by: Eike Ziller <eike.ziller@qt.io>