LanguageClientManager: stop leaking unrelated files to clients
Clients with the "requiresProject" setting were opening all kind of
files unrelated to the clients project.

It turns out that the early exit in documentOpenedForProject:
if (!project->isKnownFile(filePath))
    continue;
does not stop the code later on to open the file in the unrelated
client.
Therefore, don't append "clients" into "allClients" in case of
RequiresProject settings.
Also make clients const: we don't need to remove clients from it in the
RequiresProject case because we don't append clients into allClients
anymore.

Continue to append clients into allClients for the RequiresFile and
AlwaysOn cases.

Fixes: QTCREATORBUG-33101
Change-Id: I9748d54bc1c6e2c5c06e4b2d19989ae4268995e4
Reviewed-by: David Schulz <david.schulz@qt.io>