LanguageClientManager: Relax check for project
Relax the checks that tests whether a file is part of a project, and
also accept files inside a project source directory as being part of the
project.

This allows to run qmlls on .qml files that are part of the project but
not directly used by the project, for example some .qml files used for
testing in the qtdeclarative repo. Another use case is when users create
a new qml file and write something in it before adding it to the
project: the users might be annoyed at qmlls not running on their new file.

Change-Id: I637e6687a214c04bef72bb6e728cdb7657b93b85
Reviewed-by: David Schulz <david.schulz@qt.io>