LanguageClientManager: activate client of current configuration
Qt Creator opens newly opened files into the wrong qmlls instance. In
documentOpenedForProject, "activateDocument" is set to true and all
qmlls instances are activated for the document.

Fix the code to only activate the qmlls instance that is running for the
active build configuration, instead of activating all clients.
openDocumentWithClient() opens the document and sets the client as the
current instance, while openDocument() does not do any client activation.

Fixes a bug where qmlls is selected for the wrong build configuration by
default on a project with multiple build configurations.

Change-Id: I6de12b10a5ce5576339fae25945e19e4714a7235
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>