QmlDesigner: Handle "data" property removal of material library
In navigator view, the "data" property removal of a node is interpreted
as removal of all node's children, so added similar handling to
detect if material library or its content is deleted in this manner.

This fixes the issue of not detecting redo of deletion of all materials
and textures in material browser, or all content of the scene.

Fixes: QDS-13571
Change-Id: I0643e3714e5dbc974b6870419a451f0896aafd21
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>