Utils: Replace TextCodec in TextFileFormat::decode() by TextEncoding
The filling of partial buffers seems not possible with QStringDecoder,
and it's arguably not needed when dropping the chunked approach.

Original decoding thome was 128ms for a 110MB test file, by dropping
the chunked approach and the verification that's even down to 75ms,
dropping the chunked approach by retaining the verification by
roundtrip end up at 167ms. In any case, all these times are dwarfed
by the QTextDocument setup later, which takes 15s.

Change-Id: Ic21ab027fa385fc0af34399a11fb52740955e954
Reviewed-by: David Schulz <david.schulz@qt.io>