Macros: Add better examples for "prefix" variables
Variables registered with `registerPrefix` are shown in the variable
chooser as `Prefix:<value>`, and that was used to show the "current
value" as the example in the variable chooser.

Add an explicit "example value" to the registerPrefix calls that is used
instead of "<value>" for the expanded example in the variable chooser.

Fixes: QTCREATORBUG-33120
Change-Id: I4cc522856bf9a7fefeea2ea498de5f67d503b86a
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>