ProjectExplorer: Better separation between project managers and QmlJS
Previously, individual project managers depended on QmlJS, and
for updates they pulled a pre-filled ProjectInfo struct from
QmlJS, amended it, and pushed it back.

In the new approach, indvidual project managers can add to a similar
pre-filled structure from the buildsystem base implementation, after
which the base implementation emits a signal containing the data,
which is caught by QmlJSTool::ModelManager and there translated into
QmlJS's own ProjectInfo struct.

Advantage here is that the more of the necessary amending can
be done centrally in the new BuildSystem::updateQmlCodeModel()
and the indvidual project managers do not have to depend on
QmlJS anymore.

Change-Id: Ic3fbc55b9e3d980c55ac478bdf5672391214fa7b
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>