TextEditor: fix crash when closing editor while showing completions
The destructor of the code assistant calls back into the parent editor
widget via the destroy context functionality. If the code assistant is
destructed by the destruction of the text editor widget private this
call is not safe. Explicitly destroy the context from the editor widget
destructor before destructing any of the widget members.

Fixes: QTCREATORBUG-33079
Change-Id: I1b623ab2083283490f4f226f3e6665dd75bb0f2a
Reviewed-by: Eike Ziller <eike.ziller@qt.io>