Debugger: Use the correct guard object for connects
During asynchronous shutdown, an event loop is run and state changes can
happen, even though the main window is already hidden. We got a crash
report where the event handling results in a mode change, and that calls
DebuggerMainWindow::leaveDebugMode() with the debugger main window
already deleted by the Debugger plugin's aboutToShutdown() method.

Use the debugger main window as the guard for the connections from the
mode manager that do calls on it.

Change-Id: I196c4bd23ea3c3eaeea6b788fd22c30f61b06b03
Reviewed-by: hjk <hjk@qt.io>