macOS: Fix opening files from Finder if app is not running
Specifically when the LicenseChecker plugin is present, but also in any
other case where a plugin might run an event loop during startup (which
should generally be avoided, but anyway).

If an event loop runs before we connect to the apps fileOpenRequest, any
FileOpen events from startup get lost. So make sure that we connect
early during plugin loading by connecting in the Core plugin's
constructor.

Fixes: QTCREATORBUG-33164
Change-Id: I1d894761397631f700e716739b741e014086d44c
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>