CMakePM: Revert CMake's isExternal filtering of CMake files
Amends cd52265eb2c23e9987e5fc00a44f7bbe9bf8142c
Amends 927ee16f165160ff9933b4f6272a0b2e7defe486

CMake will mark any source file that is not part of the top source
or top binary directory as external.

Qt Creator 17.0.0 used this isExternal property of source files as
files that can be hidden from the project view.

As it turns out, there are cases like `include(../common/file.cmake)` or
`add_subdirecory(../common)` that will be treated as external but the
users would want them to be part of the project.

This commit revets the usage of CMake's isExternal, by only hiding
the files that are part of the CMAKE_PREFIX_PATH and not part of
the source or build directory.

This takes care also of the vcpkg case when CMAKE_PREFIX_PATH would
be set as "/".

Fixes: QTCREATORBUG-32095
Fixes: QTCREATORBUG-33152
Fixes: QTCREATORBUG-33163
Fixes: QTCREATORBUG-33180
Change-Id: Ifbd362b8d4a889412fff080d1459c68feb1e2d32
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>