Fix vanishing menu bar
The restart dialogs that we open for certain settings were leading to a
ICore::saveSettings after the main window was already hidden, which
implies that the menu bar is hidden, which we happily saved into the
settings.

This happened if two settings triggered a restart dialog. After the first
dialog is accepted, we schedule a "exit" call on the event loop. The
second dialog then ran another event loop that unexpectedly triggered
the "exit" - leading to that dialog and the main window being closed &
hidden. Finally, after finishing the event loop of the settings dialog we
triggered another settings save while the main window already is hidden.

Avoid this by not running another dialog if the user already agreed to
the restart. Also never schedule multiple exit() calls.

As an additional guard, we now also only save the menu bar setting if
the main window is actually still visible.

Fixes: QTCREATORBUG-33190
Fixes: QTCREATORBUG-32857
Change-Id: I876b8bfa94fe3176da660bce4c988c9988e14a54
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>