Fossil: Fix revisionSpec()
Pass "-r" when revision is specified.
In this way the revertFile() reimplementation is
the same as in VcsBaseClient, so we drop it.
For other usages of revisionSpec(), Fossil plugin has
its reimplementations provided, except for update().
So, we add a FossilClient::update() reimpl, calling the
base one, and since it looks like in this case the "-r"
arg isn't needed, we pass revision as the 1st arg to extraOptions.

Shorten the reimplementations of revisionSpec() in other plugins, too.

Change-Id: Ibbb0241b71f4451e7b4e547e58bee5f1cd1c6c0f
Reviewed-by: Orgad Shaneh <orgads@gmail.com>