Tabs: Prevent wheel scrolling
The wheel event implementation only results in a currentChanged event
which makes it impossible to identify it as a "user triggered" change.
That makes it ugly for tab bars that are 'manually managed' and want to
only react on user changes, like the editor tab bar.

This feature seems to be very Qt specific to begin with, turn it off.

Change-Id: Id05d1bfcdcfbf042ebab3884ae8fc61392b5cd07
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>