iOS: Mark tool chains as iOS in display name
The tool chains registered for iOS are not suitable for our desktop
kits, since they require the sysroot to the target platform to be
explicitly set. The /usr/bin toolchains that we detect for desktop are
in turn not usable for iOS, because they have their sysroot hardcoded to
macOS/desktop.

Change the name of the iOS toolchains to something containing "iOS" to
make that more explicit for users. We still need to keep the prefix
"Apple Clang" there, because older Qt Creator versions rely on that for
identification.

We also may not force-override the display name of tool chain bundles
for tool chains that Qt Creator autodetected. That is needed for tool
chains that are registered by the installer ("sdk-provided") because
these prominently contain the language identifier, which is not
appropriate for bundles.
Remove the language identifier for all tool chains that Qt Creator
autodetects - it is redundant because the tool chains are grouped by
language in the UI anyway.

Change-Id: I5a7c25e1a4792d879172e54dcd3a79d551269972
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>