QMLProjectManager: More streamlined tests
Simplified the integration tests and made them more unitary.
Previously the tests used the same data set both as the expected result
in qmlproject -> json conversion tests and as the source in
json  -> qmlproject conversion tests. This resulted in those
intertwined tests being unnecessarily hard to adjust.
Moreover, the following commit introduces a non-trivial filter behavior
that can potentially cause the result of the qmlproject -> json
conversion vary depending on the file types supported by QImage, which
makes the result of this conversion unsuitable to be used as the source
data anymore.

Task-number: QDS-15386
Change-Id: Idf657b168daa63cbfb1d4491832d889071cf794a
Reviewed-by: Marco Bubke <marco.bubke@qt.io>