Utils: Return QString from FilePath::relativePathFromDir()
FilePathification went a bit too far there: The result here can
only be used in connection to another "anchor" path, and almost
all users immediately converted it to a string, one way or the other.

So paddle back, and adjust callers.

For the few places that wanted to use the native separators in the
result, provide FilePath::relativeNativePathFromDir() as the OS cannot be
determined from the string alone.

Api: Utils::FilePath::relativePathFromDir() returns now QString
     instead of Utils::FilePath

Change-Id: Ide9683989930f960fa19f8e778036abde6801e89
Reviewed-by: David Schulz <david.schulz@qt.io>