CodeAssistant: fix crash on closing editor while assist runs
... in the case where a proposal is visible and the processor still
runs. For example when following a virtual function call a proposal
offers to jump to the base implementation while the code model checks
for potential overrides.

In this case only the processor was canceled in
CodeAssistantPrivate::destroyContext, but the proposal was still shown.
In the case of a closing editor this results in properly stopping the
processor in the explicit call from ~TextEditorWidget(), but the
following  implicit destruction of the code assist resulted in another
call to CodeAssistantPrivate::destroyContext still accessing the private
member of the TextEditorWidget. So we always need to make sure that the
proposal is also closed when destroying the context, not only in the
case where there is no code assist processor.

Amends 719944a15df899a6187ee13880644e6c4c7bb6bd

Change-Id: I2840e27edd5e5c5aa79d6f77c3cd7a6ec1d88afc
Reviewed-by: Eike Ziller <eike.ziller@qt.io>