ProjectExplorer: Only create TaskMark in TaskModel
If the TextMark is created in the TaskHub, then anyone who listens to
the taskAdded signal will get a reference to the TextMark as well.
This is not wanted as the lifetime of the TextMark defines the
visibility of the TextMark as well. Instead we create the TextMark only
in the copy of the TaskModel, whose lifetime corresponds to the lifetime
of the visible mark.

Fixes: QTCREATORBUG-33108
Change-Id: I4b343db58b73395dfa8b63c36b555ae703ecc459
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>