Editors: Switch the splitting logic
So far without tabs for editors, it was basically irrelevant were the
"new" split was created. It was only relevant that in the result the
focus ended up in the right / bottom one, and we chose to open the "new"
split at the left / top, which avoids switching focus and the "current
editor".

With tabs that doesn't really make sense anymore. A workflow of "open a
temporary split, do something there, close the split" breaks if closing
that split closes all the editors that were open in the tabs.

Switch it around that the new split is the right / bottom one, and the
focus moves there.

Change-Id: Ie7e4d676e16ffd78e6d4ce8cc2330c55c351c397
Reviewed-by: Christian Stenger <christian.stenger@qt.io>
Reviewed-by: David Schulz <david.schulz@qt.io>