Fix crash when connections to parsingFinished access build system
When closing a project while parsing is still running.

When the specific BuildSystem holds a member of the ParseGuard, we want
to make sure that that is destroyed before everything else, since that
sends the parsingFinished signal. Explicitly do that first in the
destructor. Otherwise, if any members are destroyed before the
ParseGuard member, accessing these members directly or indirectly from
the parsingFinished signal crashes.

Specifically this crashed the "Qt modules" part of the UsageStatistic
plugin that accesses the `additionalData` of the build system in its
anyParsingFinished handler.

Change-Id: I0e7b3003690faaf48b3b4108959c7ef49dd8e3f3
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>