Qt/Help: Fix extraction of 'best' documentation to show
In case Qt for MCUs is installed, we have duplicate documentation for
QML elements, in the MCU documentation and the "normal" Qt
documentation. We also by default only register the documentation for
"hightest registered Qt version".

When triggering context help for a project that uses a different Qt
version than the highest registered, it would open the MCU documentation
for QML elements instead of the "normal" Qt documentation.

The QtSupport plugin has code that tries to find the "best" documentation
for the requested help, i.e. the Qt documentation with the "best" fitting
version number compared to the Qt version of the project.

Since the MCU documentation has a different prefix for the documentation
URLs ("io.qt." instead of "org.qt-project.") we do not consider it "Qt
documentation" and do not extract a version number for it. The code in
QtSupport then considered documentation with no version as good as
documentation with the exact same version as the project's Qt, which
doesn't make sense. Instead, just always return the unversioned
documentation in addition to any "best versioned" documentation that we
found.

Fixes: QTCREATORBUG-33025
Change-Id: I725cf85d15769e5a2a9fdfaacf28611d98fed385
Reviewed-by: David Schulz <david.schulz@qt.io>