Editors/Tabs: Do not switch to tab with right click
It should just open the context menu.

For this we must adapt the "close editor" action in there to behave
sensible when triggered on a "suspended tab" (a tab without an editor,
just an Entry). It should close the tab like the user clicked on its
close button.

As a byproduct the patch also fixes closing a suspended document from the
Open Documents view with the context menu, which did nothing.

Change-Id: Iaac6dabff7be5fb2bf737d24655320c7b2d3affa
Reviewed-by: David Schulz <david.schulz@qt.io>