Utils: Change/"Fix" FilePath::resolvePath(QString tail)
So far, this diverted to FilePath::resolvePath(FilePath)
which completely replaces 'this' with the argument in
case the tail is absolute.

This behavior is ok for FilePath::resolvePath(FilePath),
as this allows to consider and "navigate" all accessible
paths as one big filesystem.

Using a tail QString that is absolute created an intermediate
absolute local FilePath which then replaced 'this', which
is typically not what callers expect - the function is often
used to 'filepathify' output from 'relatively locally' running
commands on devices, relative to the device root.

Change-Id: I8c2d43ff994b487d0c0f6ed1a29cc85599168d2a
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>