Axivion: Use last issue kind as default
By default architecture violations had been chosen to get
presented when opening a dashboard as it is the first
button in the list and the issues appear alphabetically
sorted.
It is hard to set up and not likely to have issues there,
so use the last button which likely are style issues.

Change-Id: I2b9ac1147b9e04c1f80c83c8a2066960cae5ed79
Reviewed-by: Daniel Hofmann <daniel.hofmann@qt.io>
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>