Editors/Tabs: Fix Close All Except Visible
Make sure that we close all tabs except the currently selected ones.

That means that we now also need to close all editors that are not
visible and all suspended tabs that do not also belong to suspended
editors. Without tabs that was not relevant, but now some non-active
tabs would stay open.

Change-Id: Id55ce8eb774416184c0ba4801486b75d5b6aac84
Reviewed-by: Christian Stenger <christian.stenger@qt.io>