TextEditor: optimize auto completing more than a single char
Currently we position the cursor after the auto completed text if it is
longer than a character. This is un intended when auto completing the
raw string delimiter since 1bce50a4c4d0e5b76480887b8498b8936f0acdff.

The code to jump over automatically inserted code now moved completely
to the TextEditorWidget, but we keep the complete skip char logic in the
auto completer code for now. The plan is to remove this in a separate
change.

Fixes: QTCREATORBUG-31901
Change-Id: I4f29ef32a15499a2a489ca1b98fae0980af57780
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>