FileSystemWatcher: Postpone directory change calls
When a set of files is being added to a directory, the internal
watcher notifies directory changes multiple times. So
FileSystemWatcher handles directory changes multiple times.
Then all previous files that are being watched, would be labeled
as changed.
This process causes quadratic calls.
If we postpone the calls for the same directory, we can prevent
some unnecessary calls.

Directory change slots are postponed for a while after the last call.

Fixes: QDS-15670
Change-Id: I73db977b77f36bb0189240ae6c1c66b083620af0
Reviewed-by: Eike Ziller <eike.ziller@qt.io>
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>