Git: Call onDone only on error
The 1st Group should never be canceled, even when
tasks executed there finish with an error.

The 2nd Group could cause the 1st one to be canceled
when diff command failed. To prevent the cancelation
of the 1st Group, change the top level workflow policy
to continueOnError. In this way, even when 2nd Group
failed, we continue executing the 1st Group.

Change-Id: Ie534564d07026e62ea10002448cedacea0ad2ede
Reviewed-by: Orgad Shaneh <orgads@gmail.com>
Reviewed-by: André Hartmann <aha_1980@gmx.de>