CppEditor: Don't try to read file contents for directories
This is apparently triggered for things like #include <QtTest>
where the file resides in Qt/.../gcc_64/include/QtTest/QtTest
and the include path contains both include and include/QtTest.

In the end FilePath::fileContents() is called, which produced
silently (harmless) empty output so far, but the remote bridge
now produces a "Error reading file ...include/QtTest: Is a directory"
warning.

Instead of not producing the warning, which might be helpful in
other cases, suppress the attempt to access the file content.

Change-Id: I033c69e65b21eaa222f062ecfb771755b02efb28
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>