CMakePM: Change the path to the temporary CMake preset probe
... from "$TEMP/QtCreator-XXXXXX/qtc-cmake-presets-XXXXXXXX" to
"<source-dir>/.qtcreator/qtc-cmake-presets-XXXXXXXX".

This CMake preset is just a simple CMake project that is then being
configured by the CMake tool and the C/C++ compilers are extracted from
the CMakeCache.txt.

I am thinking that the "<source-dir>" could be mapped as a remote path,
unlike the "$TEMP/QtCreator-XXXXX", and would allow setting up a remote
CMake preset.

Change-Id: I3e97de630920217b1a82977cc265dd0ffae3d536
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>