CMake: Fix SBOM generation for wininterrupt and cdb ext projects
The wininterrupt and qtcreatorcdbext projects use ExternalProject_Add
to build the projects for multiple architectures.

We try to find the Qt6 package in each of those builds to generate an
SBOM. But the found Qt6 architecture may not be compatible with the
architecture of the current build, which would lead to a failure to
find the package, because the ConfigVersion.cmake file checks for
architecture compatibility.

For these projects that don't link to Qt, we don't care about binary
compatibility, we only want to use the Qt SBOM feature.
Explicitly disable the version checks, so that we can always find the
Qt6 package, even if the architecture is not compatible.

We also need to ensure that we generate the SBOMs for each of these
sub-builds only if the main project has been configured to do so, so
add an explicit check for that.

To ensure the sub-build SBOMs are installed into the main build,
add additional install commands to do so.

Move the sbom setup function calls to the main part of the project,
to make it less entangled with the multi-build logic.

Fix the missing LIST_SEPARATOR option for the wininterrupt project,
so that the CMAKE_PREFIX_PATH arg is passed correctly.

Amends 0772c0058b08237213587849cc3a4e585a177684

Task-number: QTCREATORBUG-31681
Change-Id: I6b1ce111e5efafcbfe2bc4386a245a309fb0e117
Reviewed-by: Eike Ziller <eike.ziller@qt.io>