AutoTest: Allow new style logging options for QTest
QTest allows two different ways of specifying logging options.
The newer one allows to do multiple loggings to files beside
the logging to stdout.
The new way is supported since Qt5, so it is about time to
support it.
This patch enables users log the test run to one or more
files in different output formats if one uses Qt5 or later and
specifies the new logging output format as command line parameters.

This patch drops support for running and displaying results of Qt4
based Qt tests from the AutoTest plugin's side, running them and
viewing them will only be possible by the general ways of running
a project's target and looking into the application output.

Change-Id: Ic9f16d8a97882d591e5986d1cc7b593fd2e9134d
Reviewed-by: David Schulz <david.schulz@qt.io>