Welcome: Make Open Project filter for projects
In the progress of refactoring the welcome page, the Open Project button
started to behave like Open File or Project instead of Open Project.

Amends 196aee97b15234e45e4b4c64d61b479dc013f1c7

Remove unused internal methods and unify different "open project"
methods to the one used by the context menu on the Projects view.

Ensures creation of the corresponding Command in ICore for simplicity.
Otherwise we'd need to check for the command in the Welcome code, which
does not depend on ProjectExplorer which actually provides the
functionality.

Fixes: QTCREATORBUG-33259
Change-Id: I534a6061cc460519c9e4919270fbf288405f526a
Reviewed-by: André Hartmann <aha_1980@gmx.de>