QmlDesigner: Fix crash when sending custom notification
Some views such as 2D view detach when they are not visible, which can
happen if newly shown view is on same dock tab as the previous view.
Detached view cannot be directly used to send custom notification,
so we now use generic view from view manager to do so.

Fixes: QDS-15668
Change-Id: I0bccfcf09f8f2b1f244d43f3e9d82ace79512be3
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>